Implemented user registration with password encryption, login with hashed password check, and added user update, cancellation, and list functionalities.

+++ src/main/java/com/linktic/reservation_system/service/ReservationService.java
@@ -1,10 +1,12 @@
 package com.linktic.reservation_system.service;
 
+import com.linktic.reservation_system.exception.UserAlreadyExistsException;
 import com.linktic.reservation_system.model.Reservation;
 import com.linktic.reservation_system.repository.ReservationRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import java.util.List;
+import java.util.Optional;
 
 @Service
 public class ReservationService {
@@ -15,15 +17,35 @@ public class ReservationService {
         return reservationRepository.findByUserId(userId);
     }
 
+    public List<Reservation> getAllReservations() {
+        return reservationRepository.findAll();
+    }
+
     public Reservation createReservation(Reservation reservation) {
         return reservationRepository.save(reservation);
     }
 
     public Reservation updateReservation(Reservation reservation) {
-        return reservationRepository.save(reservation);
+        Optional<Reservation> existingReservation = reservationRepository.findById(reservation.getId());
+
+        if (existingReservation.isPresent()) {
+            Reservation updatedReservation = existingReservation.get();
+            updatedReservation.setReservationDetails(reservation.getReservationDetails());
+            updatedReservation.setReservationDate(reservation.getReservationDate());
+            updatedReservation.setUser(reservation.getUser());
+            return reservationRepository.save(updatedReservation);
+        } else {
+            throw new UserAlreadyExistsException("Reserva no encontrada con el ID: " + reservation.getId());
+        }
     }
 
+
     public void cancelledReservation(Long id) {
-        reservationRepository.deleteById(id);
+        Optional<Reservation> reservation = reservationRepository.findById(id);
+        if (reservation.isPresent()) {
+            reservationRepository.deleteById(id);
+        } else {
+            throw new RuntimeException("Reserva no encontrada");
+        }
     }
 }
